Keep pointer down encoder with pointer handling
